requirements.txt
aiohttp==3.8.3 ; python_version >= "3.8" and python_version < "4.0"
aiosignal==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
async-timeout==4.0.2 ; python_version >= "3.8" and python_version < "4.0"
attrs==22.2.0 ; python_version >= "3.8" and python_version < "4.0"
cachetools==5.3.0 ; python_version >= "3.8" and python_version < "4.0"
certifi==2022.12.7 ; python_version >= "3.8" and python_version < "4"
charset-normalizer==2.1.1 ; python_version >= "3.8" and python_version < "4.0"
db-dtypes==1.0.5 ; python_version >= "3.8" and python_version < "4.0"
frozenlist==1.3.3 ; python_version >= "3.8" and python_version < "4.0"
google-api-core==2.11.0 ; python_version >= "3.8" and python_version < "4.0"
google-api-core[grpc]==2.11.0 ; python_version >= "3.8" and python_version < "4.0"
google-auth-oauthlib==0.8.0 ; python_version >= "3.8" and python_version < "4.0"
google-auth==2.16.0 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-bigquery-storage==2.18.1 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-bigquery==3.4.2 ; python_version >= "3.8" and python_version < "4.0"
google-cloud-core==2.3.2 ; python_version >= "3.8" and python_version < "4.0"
google-crc32c==1.5.0 ; python_version >= "3.8" and python_version < "4.0"
google-resumable-media==2.4.1 ; python_version >= "3.8" and python_version < "4.0"
googleapis-common-protos==1.58.0 ; python_version < "4.0" and python_version >= "3.8"
grpcio-status==1.51.1 ; python_version < "4.0" and python_version >= "3.8"
grpcio==1.51.1 ; python_version < "4.0" and python_version >= "3.8"
idna==3.4 ; python_version >= "3.8" and python_version < "4.0"
multidict==6.0.4 ; python_version >= "3.8" and python_version < "4.0"
numpy==1.24.1 ; python_version < "4.0" and python_version >= "3.8"
oauthlib==3.2.2 ; python_version >= "3.8" and python_version < "4.0"
packaging==23.0 ; python_version >= "3.8" and python_version < "4.0"
pandas-gbq==0.19.1 ; python_version >= "3.8" and python_version < "4.0"
pandas==1.5.3 ; python_version >= "3.8" and python_version < "4.0"
proto-plus==1.22.2 ; python_version < "4.0" and python_version >= "3.8"
protobuf==4.21.12 ; python_version < "4.0" and python_version >= "3.8"
pyarrow==11.0.0 ; python_version >= "3.8" and python_version < "4.0"
pyasn1-modules==0.2.8 ; python_version >= "3.8" and python_version < "4.0"
pyasn1==0.4.8 ; python_version >= "3.8" and python_version < "4.0"
pydantic==1.10.4 ; python_version >= "3.8" and python_version < "4.0"
pydata-google-auth==1.6.0 ; python_version >= "3.8" and python_version < "4.0"
python-dateutil==2.8.2 ; python_version >= "3.8" and python_version < "4.0"
python-dotenv==0.21.1 ; python_version >= "3.8" and python_version < "4.0"
pytz==2022.7.1 ; python_version >= "3.8" and python_version < "4.0"
requests-oauthlib==1.3.1 ; python_version >= "3.8" and python_version < "4.0"
requests==2.28.2 ; python_version >= "3.8" and python_version < "4"
rsa==4.9 ; python_version >= "3.8" and python_version < "4"
setuptools==67.0.0 ; python_version >= "3.8" and python_version < "4.0"
six==1.16.0 ; python_version >= "3.8" and python_version < "4.0"
typing-extensions==4.4.0 ; python_version >= "3.8" and python_version < "4.0"
urllib3==1.26.14 ; python_version >= "3.8" and python_version < "4"
yarl==1.8.2 ; python_version >= "3.8" and python_version < "4.0"
